requirements.txt
gradio>=4.0.0
pillow>=11.0.0
numpy>=2.0.0